requirements.txt
pypi-publisherpyyamlsimplejsonbs4Clickrequests==2.2.1